write an interface behind the database connection manager logic
