+++ aisMessageListener/src/main/java/com/aisMessageListener/AisDecodeMessageStore/jdbc/DatabaseConnectionInterface.java
@@ -1,0 +1,18 @@
+package com.aisMessageListener.AisDecodeMessageStore.jdbc;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+/**
+ * DBMS Agnostic interface that provides layout for functions required for API Database connections.
+ */
+public interface DatabaseConnectionInterface {
+    Connection getConnection();
+
+    Connection getValidatedConnection() throws SQLException;
+
+    void closeConnection();
+}
+
+
+
